Cleanup and fix imports
