+++ src/hooks/usePlaceOrder.ts
@@ -2,8 +2,9 @@ import { useCallback } from 'react'
 import BN from 'bn.js'
 import { toast } from 'react-toastify'
 
-import { TokenDetails, PlaceOrderParams as ExchangeApiPlaceOrderParams } from 'types'
+import { TokenDetails } from 'types'
 import { exchangeApi } from 'api'
+import { PlaceOrderParams as ExchangeApiPlaceOrderParams } from 'api/exchange/ExchangeApiImpl'
 import { log } from 'utils'
 import { txOptionalParams } from 'utils/transaction'
 import { useWalletConnection } from './useWalletConnection'
