+++ src/hooks/useWalletConnection.ts
@@ -1,7 +1,8 @@
 import { walletApi } from 'api'
 import { useEffect } from 'react'
-import { WalletInfo, Command } from 'types'
+import { Command } from 'types'
 import useSafeState from './useSafeState'
+import { WalletInfo } from 'api/wallet/WalletApiImpl'
 
 export const useWalletConnection = (): WalletInfo => {
   const [walletInfo, setWalletInfo] = useSafeState<WalletInfo>({ isConnected: false })
