+++ src/services/factories/getTokenFromExchange.ts
@@ -1,9 +1,11 @@
 import Web3 from 'web3'
 
-import { Erc20Api, TokenDetails, TokenList, ExchangeApi } from 'types'
+import { TokenDetails, TokenList } from 'types'
 import { getImageUrl, log, getToken } from 'utils'
 
 import { getErc20Info } from '../helpers'
+import { Erc20Api } from 'api/erc20/Erc20ApiImpl'
+import { ExchangeApi } from 'api/exchange/ExchangeApiImpl'
 
 interface TokenFromErc20Params {
   tokenAddress: string
