+++ src/utils/exchange.ts
@@ -1,6 +1,5 @@
 import BN from 'bn.js'
-
-import { AuctionElement } from 'types'
+import { AuctionElement } from 'api/exchange/ExchangeApiImpl'
 
 const ADDRESS_WIDTH = 20 * 2
 const UINT256_WIDTH = 32 * 2
